cleaned up unused code in page.tsx for home page
