--- project/web/app/(main)/biometric-monitor/PatientComponent/AddPatientForm.tsx
@@ -1,6 +1,0 @@
-
-"use client"
-
-export default function AddPatientForm(){
-
-}
