--- project/web/app/(main)/biometric-monitor/PatientComponent/PatientProfile.tsx
@@ -1,6 +1,0 @@
-
-"use client"
-
-export default function PatientProfile(){
-
-}
