--- project/web/app/(main)/biometric-monitor/PatientComponent/PatientSelect.tsx
@@ -1,5 +1,0 @@
-"use client"
-
-export default function PatientSelect(){
-
-}
